+++ service_feedback/src/main/java/com/sistemareserva/service_feedback/client/broker/dto/UpdateRatingRequest.java
@@ -1,8 +1,10 @@
 package com.sistemareserva.service_feedback.client.broker.dto;
 
+import java.math.BigDecimal;
+
 public record UpdateRatingRequest(
     Long idQuarto,
-    Double rating
+    BigDecimal  rating
 ) {
     
 }
